remove obsolete file, small fixes
