+++ package/PartSegCore/segmentation/segmentation_algorithm.py
@@ -15,7 +15,6 @@ from ..image_operations import RadiusType
 from ..algorithm_describe_base import AlgorithmDescribeBase, AlgorithmProperty, SegmentationProfile
 from .noise_filtering import noise_filtering_dict
 from .threshold import threshold_dict, BaseThreshold, double_threshold_dict
-from .segment import close_small_holes
 
 
 class StackAlgorithm(SegmentationAlgorithm, ABC):
@@ -343,3 +342,15 @@ class AutoThresholdAlgorithm(BaseSingleThresholdAlgorithm):
 
 
 final_algorithm_list = [ThresholdAlgorithm, ThresholdFlowAlgorithm, ThresholdPreview, AutoThresholdAlgorithm]
+
+
+def close_small_holes(image, max_hole_size):
+    if image.dtype == np.bool:
+        image = image.astype(np.uint8)
+    if len(image.shape) == 2:
+        rev_conn = sitk.ConnectedComponent(sitk.BinaryNot(sitk.GetImageFromArray(image)), True)
+        return sitk.GetArrayFromImage(sitk.BinaryNot(sitk.RelabelComponent(rev_conn, max_hole_size)))
+    for layer in image:
+        rev_conn = sitk.ConnectedComponent(sitk.BinaryNot(sitk.GetImageFromArray(layer)), True)
+        layer[...] = sitk.GetArrayFromImage(sitk.BinaryNot(sitk.RelabelComponent(rev_conn, max_hole_size)))
+    return image
